add divs to Home to organise text resolve errors with CopyText on Confirm component improve "Copied" effect
